+++ packages/web/src/javascripts/Components/NoteGroupView/NoteGroupView.tsx
@@ -7,6 +7,7 @@ import MultipleSelectedFiles from '../MultipleSelectedFiles/MultipleSelectedFile
 import { ElementIds } from '@/Constants/ElementIDs'
 import FileView from '@/Components/FileView/FileView'
 import Icon from '../Icon/Icon'
+import { FileDnDContext } from '@/Components/FileDragNDropProvider/FileDragNDropProvider'
 
 type State = {
   showMultipleSelectedNotes: boolean
@@ -22,6 +23,9 @@ type Props = {
 }
 
 class NoteGroupView extends PureComponent<Props, State> {
+  static override contextType = FileDnDContext
+  declare context: React.ContextType<typeof FileDnDContext>
+
   private removeChangeObserver!: () => void
 
   constructor(props: Props) {
@@ -80,6 +84,8 @@ class NoteGroupView extends PureComponent<Props, State> {
   }
 
   override render() {
+    const fileDragNDropContext = this.context
+
     const shouldNotShowMultipleSelectedItems =
       !this.state.showMultipleSelectedNotes && !this.state.showMultipleSelectedFiles
 
@@ -113,14 +119,14 @@ class NoteGroupView extends PureComponent<Props, State> {
           )}
           {this.state.showMultipleSelectedFiles && (
             <MultipleSelectedFiles
-              application={this.application}
               filesController={this.viewControllerManager.filesController}
               selectionController={this.viewControllerManager.selectionController}
-              featuresController={this.viewControllerManager.featuresController}
-              filePreviewModalController={this.viewControllerManager.filePreviewModalController}
-              navigationController={this.viewControllerManager.navigationController}
-              notesController={this.viewControllerManager.notesController}
             />
+          )}
+          {this.viewControllerManager.navigationController.isInFilesView && fileDragNDropContext?.isDraggingFiles && (
+            <div className="absolute bottom-8 left-1/2 z-dropdown-menu -translate-x-1/2 rounded bg-info px-5 py-3 text-info-contrast shadow-main">
+              Drop your files to upload them
+            </div>
           )}
           {shouldNotShowMultipleSelectedItems && this.state.controllers.length > 0 && (
             <>